+++ src/main/java/khirtech/coronadz/demo/Errorhandler/GlobalErrorhandler.java
@@ -1,0 +1,4 @@
+package khirtech.coronadz.demo.Errorhandler;
+
+public class GlobalErrorhandler {
+}
