+++ src/main/java/khirtech/coronadz/demo/infection/InfectionRepository.java
@@ -1,4 +1,4 @@
-package khirtech.coronadz.demo.repositories;
+package khirtech.coronadz.demo.infection;
 
 import khirtech.coronadz.demo.entities.Infection;
 import org.springframework.data.mongodb.repository.MongoRepository;
