+++ src/main/java/khirtech/coronadz/demo/infection/InfectionService.java
@@ -1,0 +1,7 @@
+package khirtech.coronadz.demo.infection;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class InfectionService {
+}
