+++ src/main/java/khirtech/coronadz/demo/utils/validators/UserValidator.java
@@ -1,0 +1,4 @@
+package khirtech.coronadz.demo.utils.validators;
+
+public class UserValidator {
+}
